Ensure that DB access is the only way to generate a website (ie. we don't go back to the original files).
Previously, our Photo class was attempting to switch between the DB and
the file as necessary to read data from it. I think it's preferable to
read from the files only when writing to the DB intially, and let all
access in future come from the DB. This should be faster, as well as
making it clear where the data comes from at all times.